home/requirements.txt
dominate
git+https://github.com/Knio/whirl.git


requests
python-kasa
IPython
pyserial
pynput
toml
